+++ Assets/Animation/Canvas.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 995a357d1552c0b46a4d1915b28f71b2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
