+++ Assets/Animation/GameOverScene.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a1bd84b4ead72f04d8f3f169f3009c6b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
