+++ Assets/Fonts/SF Distant.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4c894a014d793894eb12cb8633a9a9a2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
