+++ Assets/Fonts/SF Distant/SF Distant Galaxy AltOutline Italic.ttf.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: 566f827e1f413854798e1fcc2b4eb061
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontName: SF Distant Galaxy AltOutline
+  fontNames:
+  - SF Distant Galaxy AltOutline
+  fallbackFontReferences:
+  - {fileID: 12800000, guid: 41001f669e01d884e852426bf798fc6f, type: 3}
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
