+++ Assets/Fonts/SF Distant/SF Distant Galaxy AltOutline.ttf.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 41001f669e01d884e852426bf798fc6f
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontName: SF Distant Galaxy AltOutline
+  fontNames:
+  - SF Distant Galaxy AltOutline
+  fallbackFontReferences: []
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
