+++ Assets/Fonts/SF Distant/SF Distant Galaxy Italic.ttf.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: cc65cbdc45611bb4889c50172587bb0d
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontName: SF Distant Galaxy
+  fontNames:
+  - SF Distant Galaxy
+  fallbackFontReferences:
+  - {fileID: 12800000, guid: 0435c2ce2f8c4f14eaa745b2360c6237, type: 3}
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
