+++ Assets/Fonts/SF Distant/SF Distant Galaxy Outline Italic.ttf.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 80e46c074cb788842963c0c2cccc7823
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontName: SF Distant Galaxy Outline
+  fontNames:
+  - SF Distant Galaxy Outline
+  fallbackFontReferences: []
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
