+++ Assets/Fonts/SF Distant/SF Distant Galaxy Symbols Italic.ttf.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 1c8cbec6d04395942a9bf4bc5138d885
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontName: SF Distant Galaxy Symbols
+  fontNames:
+  - SF Distant Galaxy Symbols
+  fallbackFontReferences: []
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
