+++ Assets/Prefabs/Enemies/Asteroid Type 1 Enemy.prefab
@@ -29,7 +29,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2664397914477520091}
   m_LocalRotation: {x: 0, y: 1, z: 0, w: -0.00000004371139}
-  m_LocalPosition: {x: 1.844487, y: -1.5999999, z: 278.8606}
+  m_LocalPosition: {x: 1.8445746, y: -1.5999999, z: 279.0219}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 7877099743686574080}
