+++ Assets/Prefabs/Powerups/Heal Powerup.prefab
@@ -26,7 +26,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 9103886400152068252}
   m_LocalRotation: {x: -0.86602545, y: 0.49999997, z: 0.000000037855173, w: -0.000000021855692}
-  m_LocalPosition: {x: 2.4, y: 3, z: 278.8606}
+  m_LocalPosition: {x: 2.4, y: 3, z: 279.0219}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 7210035535302636612}
