+++ Assets/Prefabs/Powerups/Megabomb Powerup.prefab
@@ -25,8 +25,8 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 8082590533902896425}
-  m_LocalRotation: {x: 0.86602545, y: 0.49999997, z: -0.000000037855173, w: -0.000000021855692}
-  m_LocalPosition: {x: -2.4, y: 3, z: 278.8606}
+  m_LocalRotation: {x: 0, y: 1, z: 0, w: -0.00000004371139}
+  m_LocalPosition: {x: 1.8445746, y: -1.5999999, z: 279.0219}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 667277020090645183}
