--- Assets/Scenes.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: 137fe1b78810a81498d835be344f2d69
-folderAsset: yes
-DefaultImporter:
-  userData:
