--- Assets/Scenes/Controls.unity.meta
@@ -1,4 +1,0 @@
-fileFormatVersion: 2
-guid: e31151469f9db41bb99384b804f43c3d
-DefaultImporter:
-  userData:
