--- Assets/Scenes/Drag And Drop.unity.meta
@@ -1,4 +1,0 @@
-fileFormatVersion: 2
-guid: 40116c4af15e1444f8e21d1de1f4678f
-DefaultImporter:
-  userData:
