--- Assets/Scenes/Draggable Panel.unity.meta
@@ -1,4 +1,0 @@
-fileFormatVersion: 2
-guid: 838436750d0cf48a685923371f66cea6
-DefaultImporter:
-  userData:
