--- Assets/Scenes/Layout Groups.unity.meta
@@ -1,4 +1,0 @@
-fileFormatVersion: 2
-guid: 13b46e70f240943ccb1f2665a779f3a6
-DefaultImporter:
-  userData:
