--- Assets/Scenes/Lighting.unity.meta
@@ -1,4 +1,0 @@
-fileFormatVersion: 2
-guid: f7590198924b11e47b1b33cdd7ec91ac
-DefaultImporter:
-  userData:
