--- Assets/Scenes/RenderTexture.unity.meta
@@ -1,4 +1,0 @@
-fileFormatVersion: 2
-guid: d5545059c6381d848ab3d612c68f2669
-DefaultImporter:
-  userData:
