+++ Assets/Scripts/Managers/GameOver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac1490ddc5e46134d83b078e19046d08
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
