+++ Assets/Scripts/Managers/ScoreManager.cs
@@ -14,6 +14,7 @@ public class ScoreManager : MonoBehaviour
     {
         text = GetComponent<Text>();
         score = 0;
+        DontDestroyOnLoad(this);
     }
