+++ Assets/_Scenes/GameOver.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8d1beb6a23ebd2b4e8130e04ff1e9aff
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
